requirements.txt
line-bot-sdk
flask
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymysql